test(auth): run the account-creation assertions in their own process, against real libsql (#875)
